requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
braintree==4.15.2
Brotli==1.0.9
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.6.0
Deprecated==1.2.13
Django==4.0.5
django-admin-sortable2==2.0.4
django-js-asset==2.0.0
django-mptt==0.13.4
easy-thumbnails==2.8.1
fonttools==4.33.3
html5lib==1.1
idna==3.3
kombu==5.2.4
lxml==4.9.0
packaging==21.3
Pillow==9.1.1
pkg_resources==0.0.0
prompt-toolkit==3.0.29
pycparser==2.21
pydyf==0.2.0
pyparsing==3.0.9
pyphen==0.12.0
pytz==2022.1
redis==4.3.3
reportlab==3.6.10
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
weasyprint==55.0
webencodings==0.5.1
wrapt==1.14.1
zopfli==0.2.1
